Add a complex Chronos example: ETL with Spark
This commit introduces a more complex Chronos example, where a single
parent job runs on a schedule, and two additional dependency-based jobs
only run if their parent was successful.

Synopsis:
This example reads in Leo Tolstoy's "War and Peace" from Project
Gutenberg and saves it as a text file in HDFS. A Spark job then reads
the file from HDFS and counts the number of times each word appears in
the text. This dataset is then stored on HDFS for later consumption.
Finally, a third job runs `cat` against the `part-*` files in HDFS and
sorts them in reverse order, displaying the top 20 words in the book.