aurora docker-run example with nginx
This commit introduces an example Aurora configuration for launching a
Docker container, specifically the Nginx 1.9 image. Note that this
example "shells out" to the `docker run` command, because Aurora 0.9.0
does not yet expose all of the DockerInfo fields present in Mesos.
Because of this, we're limited in our ability to specify arbitrary
options, including port mappings and volumes, from within the Aurora
DSL.