add a build script for apache aurora
Some of the feedback for the final manuscript review indicated that
there was too much detail in the build/install instructions for Apache
Aurora. Since Apache doesn't generally distribute packages, there aren't
official packages for Aurora 0.9.0 (although there are some packaging
efforts in the works).

This commit introduces `aurora-build.sh`, which will allow the reader to
build all of the various Aurora components for RHEL / CentOS 7 and
Ubuntu 14.04. This should cut down on the printed instructions in the
final manuscript, and will allow me to make changes to the build script
in the future (if needed) without issuing an update / errata to the book.